refactor(tools): split monolithic tools.ts into per-resource modules
Break the 1577-line src/tools.ts into a thin orchestrator plus one module
per resource under src/tools/, with a shared runtime:

- src/tools/runtime.ts: ToolContext + the per-server runtime (caller-tracking,
  authenticated REST/Serverless/GraphQL clients, the v1/v2 backend resolver,
  jsonReply, annotation presets) built once and threaded into each registrar.
- src/tools/{catalog,pods,endpoints,jobs,templates,network-volumes,registries}.ts:
  the tools for each resource.
- src/tools.ts: builds the runtime and calls each registrar; re-exports
  ToolContext so entrypoints/tests keep their import path.

No behavior change: identical 39-tool surface, byte-identical outbound requests
(the golden wire tests are unchanged and green). Adds a frozen full-surface
snapshot test so a future refactor or dropped registrar import can't silently
drop a tool.
